requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.42.0
google-genai>=1.0.0
python-dotenv==1.0.1
python-docx==1.1.2
python-multipart==0.0.20
